More appropriate format

+++ libraries/device/general/ic9100_lib.h
@@ -83,7 +83,7 @@ struct ic9100_handle
     uint8_t swrmeter;
     uint8_t alcmeter;
     uint8_t compmeter;
-    uint16_t opmode;
+    uint8_t opmode;
     float bandpass;
     float powerin;
     float powerout;
